🚚 Rename and move neural worker procedures definition
